Restructured calibration module and improved docs. (#166)
* * Restructured calibration module and improved docs.

* * docs.

* * Fixed docstring error.
* Improved structure of the docs.

* * Split spectroscopy and its analysis class.

* * Added some intro text to the data processor.

* * Doc and lint.

* * Black

* * Moved the CalibrationError to the package exceptions.py
