+++ docs/apidocs/calibration.rst
@@ -1,0 +1,6 @@
+.. _qiskit-experiments-calibration:
+
+.. automodule:: qiskit_experiments.calibration
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
