+++ qiskit_experiments/calibration/drag.py
@@ -23,8 +23,8 @@ import qiskit.pulse as pulse
 from qiskit.providers.options import Options
 
 from qiskit_experiments.base_experiment import BaseExperiment
-from qiskit_experiments.calibration.exceptions import CalibrationError
-from qiskit_experiments.calibration.experiments.drag_analysis import DragCalAnalysis
+from qiskit_experiments.exceptions import CalibrationError
+from qiskit_experiments.calibration.analysis.drag_analysis import DragCalAnalysis
 
 
 class DragCal(BaseExperiment):
