+++ qiskit_experiments/exceptions.py
@@ -17,3 +17,7 @@ from qiskit.exceptions import QiskitError
 
 class AnalysisError(QiskitError):
     """Class for errors raised by experiment analysis."""
+
+
+class CalibrationError(QiskitError):
+    """Errors raised by the calibration module."""
